release 34.33
Changed the missing style stones message to help prevent false bug reports.
Added an output line when a Master Writ was prevented from being accepted, making it clear that this is an optional feature of the addon.
Added some comments to QuestHandler.lua
Expanded the Quick Quest Dialogue to work for Master Writs.
When crafting a master writ, the addon will tell you what crafting type it is for.