release 35.0
The item withdrawal from bank has been refactored. Note that this may result in bugs, but it should work fine. However, it should support German and French. If you encounter any issues where an item is not withdrawn, please take a screenshot of the quest in your journal, and also copy and paste the item link. I will need this info to fix it.
German Master Writs will now craft properly.
Fixed a lua error that would show up in the Japanese version, that was caused by a missing string. (There is now a placeholder)