release 35.2
Fixed an issue where alchemy crafting mats would not be withdrawn. Note that if you have been having issues with many other items, this will probably not fix it.
Added a string that was missing in the Settings Menu.
The addon will now remove items that fit quest requirements for all crafting writs. This includes Glyphs, Weapons and Armour, in case you inadvertently deposit some.