release 36.18
Fixed a lua error when crafting master writs. When a quest was updated, the addon removed all craft requests associated, then attempted to re-add everything if necessary. This helped deal with deconning stuff, but created an issue with because the quest update event ran before the craft complete event, messing with the queue.