release 36.20
Added more style stone error messages. Now, if the user does not have many styles known, the addon will direct the user to learn more styles, or if not many styles are allowed in the settings menu, it will direct them to it. This should cut down on the 'bug' reports I get based on style stones.
Added more support for Japanese Enchanting Master Writs.
Fixed the issue with German Prachtleder writs.
The addon will now withdraw potions in German when the quest uses Tränke instead of Trank.